rename(web): 서버 전용 헬퍼 위치이동
